+++ lib/widgets/form_elements.dart
@@ -4,26 +4,31 @@ class CheckboxWithLabel extends StatelessWidget {
   final bool value;
   final ValueChanged<bool?>? onChanged;
   final Widget label;
+  final EdgeInsetsGeometry? padding;
 
   const CheckboxWithLabel({
     Key? key,
     required this.value,
     required this.label,
     required this.onChanged,
+    this.padding = const EdgeInsets.only(right: 8),
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    Widget child = Row(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        Checkbox(value: value, onChanged: onChanged),
+        label,
+      ],
+    );
+    if (padding != null) {
+      child = Padding(padding: padding!, child: child);
+    }
     return InkWell(
       onTap: onChanged == null ? null : () => onChanged!(!value),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Checkbox(value: value, onChanged: onChanged),
-          label,
-          const SizedBox(width: 8),
-        ],
-      ),
+      child: child,
     );
   }
 }
@@ -33,6 +38,7 @@ class RadioWithLabel<T> extends StatelessWidget {
   final T? groupValue;
   final ValueChanged<T?>? onChanged;
   final Widget label;
+  final EdgeInsetsGeometry? padding;
 
   const RadioWithLabel({
     Key? key,
@@ -40,23 +46,28 @@ class RadioWithLabel<T> extends StatelessWidget {
     required this.groupValue,
     required this.label,
     required this.onChanged,
+    this.padding = const EdgeInsets.only(right: 8),
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    Widget child = Row(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        Radio(
+          value: value,
+          groupValue: groupValue,
+          onChanged: onChanged,
+        ),
+        label,
+      ],
+    );
+    if (padding != null) {
+      child = Padding(padding: padding!, child: child);
+    }
     return InkWell(
       onTap: onChanged == null ? null : () => onChanged!(value),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Radio(
-            value: value,
-            groupValue: groupValue,
-            onChanged: onChanged,
-          ),
-          label,
-        ],
-      ),
+      child: child,
     );
   }
 }