+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.5.7
+- buff/effect filter
+- support new(after 6th) and legacy drop rate data
+- summon algorithm changed: statistic r5 prob 1.44%->1.24%->1.04%
+
 ## 1.5.6
 - add append skill in plan list page
 - add account in homepage